+++ SrcPony/at90sxx.h
@@ -60,7 +60,7 @@ class At90sxx : public Device
 
   protected:
 
-	int QueryType(long &type);
+	int QueryType(quint32 &type);
 
   private:
